docs: 40 docs positioning workshop (#41)
<!-- Thanks for your contribution! As part of our Community Growers
initiative 🌱, we're donating Justdiggit bunds in your name to reforest
sub-Saharan Africa. To claim your Community Growers certificate, please
contact David Berenstein in our Slack community or fill in this form
https://tally.so/r/n9XrxK once your PR has been merged. -->

# Description

I updated the readme to align with https://github.com/argilla-io/argilla

Closes #40 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

N.A.

**How Has This Been Tested**

N.A.

**Checklist**

N.A.