feat: add score to span suggestions (#56)
# Description

This PR add a small addition to value items specified for responses and
suggestions adding `score` field.

As we now the `check_response` implementation only validates so still
any field specified in the value for responses and suggestions will be
inserted in the database. Because of this we have some caveats:
* `score` using an integer like `1` will be store as `1` without being
transformed to float.
* When `score` is not specified we will not store inside the JSON column
`{..., "score": None}` instead `score` attribute will not be available.
* It's possible to specify `score` for responses not only suggestions.

Closes #55 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Adding new tests.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)