refactor: improve how response and suggestion value schemas are defined (#62)
# Description

This PR include some changes to improve how response and suggestion
value schemas are parsed and implemented. Before this we were not doing
any parsing at all due to the complexity that it involves.

Some additional improvements that we can do:
* Instead of using `int` or `str` for some of our types specify
`StrictStr` and `StrictInt` so we can avoid edge cases where `True` is
parsed as `1` and consumed by `RatingQuestionResponseValue` for example.
* Right now Pydantic errors when a value doesn't match the union of
possible values is big and difficult to understand, we can explore
improve this doing a root validation and giving a common (more simple)
message.

Closes #61 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- [x] Adding more tests and refactoring some of the existing ones.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)