feat: suggestion score supporting multiple scores (#98)
# Description

This PR add the necessary changes so now suggestions allow a single
score value or a list of score values for `score` attribute.

This change has the following restrictions:
* `score` is optional (as before).
* `score` can be a score value in the range `[0, 1]` (as before).
* `score` can be a non empty list of score values in the range `[0, 1]`
(added change).
* when `score` is a list of scores values is not possible to specify
`None` as a score value, so we are losing the optionality of scores when
defined as a list (added change).
* when `score` is a list of scores values the number of items inside the
list should match with the number of items in the `value` attribute
(added change).

> [!NOTE]
> I will create another PR removing `score` attribute for suggestions
linked to span questions for individual value items.

Closes #93

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [ ] Test A
- [ ] Test B

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)