feat: suggestions multi score downgrade implementation (#112)
# Description

This PR includes changes implementing the missing rollback/downgrade for
migration updating `score` column for suggestion to be of type JSON.

The discussed logic was the following:
1. Preserve values that have a single float value. 
2. Discard anything else setting it as `NULL` (so list of floats are
downgraded and set to `NULL`).

I have tested the migration successfully with SQLite and PostgreSQL.

Closes #113 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Running the rollback of the migration using SQLite and PostgreSQL
and validating that the values are the expected ones.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)