feat: expose Hugging Face spaces environment variables on `/api/_status` endpoint (#121)
# Description

This PR add changes to `/api/_status` endpoint exposing Hugging Face
spaces environment variables only when a `SPACE_ID` environment variable
is available.

The response of the endpoint will gain a new `huggingface` attribute
with the following structure:

```json
{
    "space_id": "space-id",
    "space_title": "space-title",
    "space_subdomain": "space-subdomain",
    "space_host": "space-host",
    "space_repo_name": "space-repo-name",
    "space_author_name": "space-author-name",
    "space_persistant_storage_enabled": true
}
```

If `SPACE_ID` environment variable is not available so we are not
running Argilla inside a Hugging Face space then the `huggingface`
attribute inside the response will not be available.

> [!NOTE]
> I will add an additional PR including a new setting so it's possible
to disable the warning when persistant storage is disabled.

Closes #120 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [ ] Add new tests to our suite.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)