feat: add new environment variable to disable Hugging Face persistant storage warning (#124)
# Description

This PR add two new changes:
* A new environment variable
`ARGILLA_SHOW_HUGGINGFACE_SPACE_PERSISTANT_STORAGE_WARNING` with a
default `True` value, allowing users to set it on Hugging Face Spaces to
disable the modal warning even when the persistant storage is disabled.
* Changes on `/api/_status` endpoint with a new `argilla` attribute that
will expose the value of the environment variable with
`{"show_huggingface_space_persistant_store_warning": true}`. The
attribute will only be present if Argilla is running on a Hugging Face
space.

So @damianpumar in the frontend will take the value of this attribute
and the value added in the previous PR
(`space_persistant_storage_enabled`) and show the warning when the
endpoint response attributes are as follow:

```json
{
  "argilla": {
    "show_huggingface_space_persistant_storage_warning": true
  },
  "huggingface": {
    "space_persistant_storage_enabled": false,
  }
}
```

Closes #123 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Adding new tests.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)