feat: set suggestions score to null when migrating score to JSON and value attribute is an array (#129)
# Description

This PR improves the database migration changing `score` column from
`suggestions` table to be JSON adding an additional update so `score`
values associated to suggestions with `value` using an array are set to
`NULL`.

This update is necessary because we don't want to support suggestions
with `value` having an array and only one float in the `score` column
(when they should have multiple `score` values).

Closes #128

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Manually testing the migration using SQLite.
- [x] Manually testing the migration using PostgreSQL. 

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)