feat: remove score attribute for suggestions associated to span questions (#101)
# Description

This PR remove the support to use `score` attribute inside value items
for creating suggestions associated to a span question.

If the value has been added prior to this change it should be included
in the suggestion response.

> [!NOTE]
> A possible nice addition could be to do a data migration where score
values for span suggestions are copied to `score` attribute as a list.
The problem with this alternative is that we should fill with some value
the `scores` that have not been assigned because `score` inside span
suggestions was optional.
> We can discuss this.

Closes #99

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Adapting tests to new change

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)