feat: add support to new `options_order` settings attribute for multi label selection questions (#133)
# Description

This PR add changes so a new attribute with name `options_order` can be
set for multi selection question settings.

This attribute is an enumerable that can have the following values:
* `natural`: the default value, showing that the options/labels should
be ordered in their natural order where they have been specified for the
questions.
* `suggestion`: showing that options/labels should be ordered giving
priority to the labels that have suggestions (having or not an assigned
`score`) and sorting by `score` when available.

Some more details about this implementation:
- When retrieving a multi label selection question without
`options_order` in his settings, `options_order: natural` will be
returned inside the settings.
- When creating a multi label selection question if `options_order` has
not been defined as part of the creation body it will be inserted as
`options_order: natural` inside the database. Notice how a `None` value
is not accepted.
- When updating a multi label selection question `options_order` is
optional but `None` is not a valid value.

Closes #132 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] Adding new tests and improving old ones.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)