feat: add new endpoint to create user tokens on API v1 (#138)
# Description

This PR implements a new endpoint on API v1 to create tokens for users.
It works the same than the one in API v0 with small changes.

The new endpoint created is `POST /api/v1/token`. I have used `/token`
instead of `/login` because I think we are creating tokens and it's
better if we continue using that naming for it.

@frascuchon tell me if we still have a good reason for rename it to
`/login` and I will change it if necessary.

Closes #137 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- [x] Adding new tests.
- [x] I have check that old tests are passing using the new endpoint
too.

**Checklist**

- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [ ] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I filled out [the contributor form](https://tally.so/r/n9XrxK)
(see text above)
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)